+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   * Fixed some CodeActions providing incorrect fix edits
 * Sumneko plugin:
   * Include `__modname__` tag when ingoring commands at start of line
+* VSCode:
+  * Migrate Mod Portal API key from VSCode key storage to `keytar` for consistent behavior across packages
 
 ## 1.1.28
 * Debugger: